tetragon: Move mbset related code in bpf_mbset.h header
Trying to keep mbset code in one place wrt upcoming changes.

We still have few mbset bits in bpf_execve_map_update.c and
it seems better to keep it there for now.

Also removing stale comment about supporting only In operator.
