+++ bpf/process/bpf_execve_event.c
@@ -11,6 +11,7 @@
 #include "bpf_helpers.h"
 #include "bpf_rate.h"
 #include "bpf_errmetrics.h"
+#include "bpf_mbset.h"
 
 #include "policy_filter.h"
 
@@ -41,17 +42,6 @@ struct {
 	__type(key, __u32);
 	__type(value, struct msg_data);
 } data_heap SEC(".maps");
-
-/* tg_mbset_map holds a mapping from (binary) paths to a bitset of ids that it matches. The map is
- * written by user-space and read in the exec hook to determine the bitset of ids of a binary that
- * is executed.
- */
-struct {
-	__uint(type, BPF_MAP_TYPE_HASH);
-	__uint(max_entries, 1024);
-	__type(key, __u8[MATCH_BINARIES_PATH_MAX_LENGTH]);
-	__type(value, mbset_t);
-} tg_mbset_map SEC(".maps");
 
 FUNC_INLINE __u32
 read_args(void *ctx, struct msg_execve_event *event)
@@ -264,29 +254,6 @@ execve_rate(void *ctx __arg_ctx)
 	if (cgroup_rate(ctx, &msg->kube, msg->common.ktime))
 		tail_call(ctx, &execve_calls, 1);
 	return 0;
-}
-
-/* update bitset mark */
-FUNC_INLINE
-void update_mb_bitset(struct binary *bin)
-{
-	__u64 *bitsetp;
-	struct execve_map_value *parent;
-
-	parent = event_find_parent();
-	if (parent) {
-		/* ->mb_bitset is used to track matchBinary matches to children (followChildren), so
-		 * here we propagate the parent value to the child.
-		 */
-		lock_or(&bin->mb_bitset, parent->bin.mb_bitset);
-	}
-
-	/* check the map and see if the binary path matches a binary
-	 * NB: only the In operator is supported for now
-	 */
-	bitsetp = map_lookup_elem(&tg_mbset_map, bin->path);
-	if (bitsetp)
-		lock_or(&bin->mb_bitset, *bitsetp);
 }
 
 /**
